+++ docs/index.d.ts
@@ -15,14 +15,35 @@ import {
 } from 'styled-components';
 import moment from 'moment';
 
-interface ColumnProps {
-  flex: string;
+interface ContainerProps {
+  className?: string,
+  columns?: number | string,
+  gap?: string,
+  columnGap?: string,
+  rowGap?: string,
+  height?: string,
+  minRowHeight?: string,
+  flow?: string,
+  rows?: number | string,
+  areas?: Array<string>,
+  justifyContent?: string,
+  alignContent?: string
+}
+
+interface CellProps {
+  className?: string;
+  width?: number;
+  height?: number;
+  top?: number | string;
+  left?: number | string;
+  middle?: boolean;
+  center?: boolean;
+  area?: string;
 }
 
 interface IGrid {
-  Column: StyledComponent<'div', Theme, ColumnProps>;
-  Row: StyledComponent<'div', Theme>;
-  Container: StyledComponent<'div', Theme>;
+  Container: StyledComponent<'div', Theme, ContainerProps>;
+  Cell: StyledComponent<'div', Theme, CellProps>;
 }
 
 export declare const Grid: IGrid;
@@ -150,16 +171,17 @@ interface TableColumn$RenderProps<T extends TableRowData, V = any> {
   rowValues: T;
 }
 
-export interface RowAction {
-  icon: string;
-  onClick: (id: number | string) => any;
+export interface RowAction<T extends TableRowData> {
+  icon?: string;
+  onClick?: (id: number | string) => any;
+  render?: (props: { row: T }) => ReactNode;
 }
 
 interface RowActionsRender<T extends TableRowData> {
   (prop: { row: T }): ReactNode;
 }
 
-export type RowActions<T extends TableRowData> = Array<RowAction> | RowActionsRender<T>;
+export type RowActions<T extends TableRowData> = Array<RowAction<T>> | RowActionsRender<T>;
 
 export interface TableOptions<T extends TableRowData> {
   rowSelection: {
@@ -262,16 +284,16 @@ export interface ButtonContentProps {
   iconModifiers?: Array<string>;
 }
 
-export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement>, ButtonContentProps {
+export interface ButtonProps<T> extends ButtonHTMLAttributes<HTMLButtonElement>, ButtonContentProps {
   buttonModifiers?: Array<string>;
   size?: ButtonSize;
   style?: CSSProperties;
   className?: string;
-  as?: ReactElement<HTMLAttributes<any>>;
+  as?: ComponentType<T>;
+  fullWidth?: boolean;
 }
 
-export declare class Button<T = {}> extends Component<T & ButtonProps> {
-}
+export declare class Button<T = {}> extends Component<T & ButtonProps<T>> { }
 
 export interface CardProps {
   title?: string;