+++ src/blocks/Grid/Container.js
@@ -1,19 +1,24 @@
 import styled from 'styled-components';
-import { applyStyleModifiers } from 'styled-components-modifiers';
-import { positionRelative } from 'modifiers';
-import { gridScale } from 'utils';
 
-const modifiers = {
-  fluid: () => `padding: 0;`,
-  positionRelative,
-};
+const autoRows = ({ minRowHeight = '20px' }) => `minmax(${minRowHeight}, auto)`;
 
-const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  padding: ${gridScale(0.5)};
-  ${applyStyleModifiers(modifiers)};
+const frGetter = value => (typeof value === 'number' ? `repeat(${value}, 1fr)` : value);
+
+const formatAreas = areas => areas.map(area => `"${area}"`).join(' ');
+
+const GridContainer = styled.div`
+  display: grid;
+  height: ${({ height = 'auto' }) => height};
+  grid-auto-flow: ${({ flow = 'row' }) => flow};
+  grid-auto-rows: ${autoRows};
+  ${({ rows }) => rows && `grid-template-rows: ${frGetter(rows)}`};
+  grid-template-columns: ${({ columns = 12 }) => frGetter(columns)};
+  grid-gap: ${({ gap = '8px' }) => gap};
+  ${({ columnGap }) => columnGap && `column-gap: ${columnGap}`};
+  ${({ rowGap }) => rowGap && `row-gap: ${rowGap}`};
+  ${({ areas }) => areas && `grid-template-areas: ${formatAreas(areas)}`};
+  ${({ justifyContent }) => justifyContent && `justify-content: ${justifyContent}`};
+  ${({ alignContent }) => alignContent && `align-content: ${alignContent}`};
 `;
 
-/** @component */
-export default Container;
+export default GridContainer;